Create R_file_sentiment_analysis
This is an R code for sentiment analysis of the comment section under any YouTube video. I wrote this code for a recent analysis project. You can use it if you like. I've added comments to make it easy to use and understand, even for beginners. You can just add your own YouTube API key and YouTube video ID in the defined placeholders.